# openapi.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/comments/task/{taskId}": {
      "get": {
        "tags": [
          "Комментарии"
        ],
        "summary": "Получить список комментариев по id задачи",
        "operationId": "getCommentsByTaskId",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список комментариев получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Комментарии"
                }
              }
            }
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Задачи с таким id не найдено"
          }
        }
      },
      "post": {
        "tags": [
          "Комментарии"
        ],
        "summary": "Добавить комментарий к задаче",
        "operationId": "createComment",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Комментарий добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Комментарий"
                }
              }
            }
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Задачи с таким id не найдено"
          }
        }
      }
    },
    "/auth/sign-up": {
      "post": {
        "tags": [
          "Регистрация"
        ],
        "summary": "Регистрация пользователя",
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Данные для регистрации"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Пользователь зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JWT токен"
                }
              }
            }
          },
          "400": {
            "description": "Не валидный email или с таким уже зарегистрировались"
          }
        }
      }
    },
    "/auth/sign-in": {
      "post": {
        "tags": [
          "Авторизация"
        ],
        "summary": "Авторизация пользователя",
        "operationId": "signIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Данные для авторизации"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Пользователь авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JWT токен"
                }
              }
            }
          },
          "400": {
            "description": "Невалидный email или неверные логин или пароль"
          },
          "401": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
...
    "/comments/{id}": {
      "delete": {
        "tags": [
          "Комментарии"
        ],
        "summary": "Удалить комментарий",
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Комментарий удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Комментарий"
                }
              }
            }
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Комментария с таким id не найдено"
          }
        }
      },
      "patch": {
        "tags": [
          "Комментарии"
        ],
        "summary": "Редактировать комментарий",
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Комментарий отредактирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Комментарий"
                }
              }
            }
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Комментария с таким id не найдено"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Задачи"
        ],
        "summary": "Получить задачу",
        "operationId": "getTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Задача получена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Задача"
                }
              }
            }
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Задачи с таким id не найдено"
          }
        }
      },
      "delete": {
        "tags": [
          "Задачи"
        ],
        "summary": "Удалить задачу",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Задача обновлена"
          },
          "403": {
            "description": "Не хватает прав"
          },
          "404": {
            "description": "Задачи с таким id не найдено"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Задача": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "id задачи"
          },
          "title": {
            "type": "string",
            "description": "Заголовок задачи"
          },
          "description": {
            "type": "string",
            "description": "Описание задачи"
          },
          "status": {
            "type": "string",
            "description": "Статус задачи",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "priority": {
            "type": "string",
            "description": "Приоритет задачи",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Временная метка создания задачи"
          },
          "author": {
            "$ref": "#/components/schemas/Пользователь",
            "description": "Автор"
          },
          "performer": {
            "$ref": "#/components/schemas/Пользователь",
            "description": "Исполнитель"
          }
        }
      },
      "Пользователь": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "id пользователя"
          },
          "username": {
            "type": "string",
            "description": "Логин (почта)"
          },
          "role": {
            "type": "string",
            "description": "Роль пользователя",
            "enum": [
              "ROLE_ADMIN",
              "ROLE_USER"
            ]
          }
        }
      },
...
      "Комментарий": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "id комментария"
          },
          "comment": {
            "type": "string",
            "description": "Текст комментария"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "Временная метка создания комментария"
          },
          "author": {
            "$ref": "#/components/schemas/Пользователь",
            "description": "Автор комментария"
          }
        }
      },
...
      "Комментарии": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32",
            "description": "Количество комментариев"
          },
          "comments": {
            "type": "array",
            "description": "Комментарии",
            "items": {
              "$ref": "#/components/schemas/Комментарий"
            }
          }
        }
      }
    }
  },
  "info": {
    "description": "TaskManagementSystem openapi"
  }
}